requirements.txt
torch
transformers
pytorch_lightning
evaluate
Pillow
matplotlib
